Add files via upload
This is a Heart diseases detection system . which help people to predict there heart condition, As all of us know that the heart attack problem is increasing day by day. and in every age group so it is necessary to regularly check our heart condition and predict the condition of heart as well as about the heart attack rate . And get the treatment at right time. So, this system helps to do this task very well.